// src/common/type.ts
export interface topicDataType {
  author: {
    loginname: string;
    avatar_url: string;
  };
  author_id: string;
  created_at: string;
  good: boolean;
  id: string;
  last_reply_at: string;
  reply_count: number;
  tab: string;
  title: string;
  top: boolean;
  visit_count: number;
}